pyproject.toml
[project]
name = "fashion-multimodal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.8.2",
    "faiss-cpu>=1.13.2",
    "gradio>=6.9.0",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
]

[tool.uv.sources]
# Pins these packages to the custom index defined below
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true  # Prevents other packages from accidentally using this index

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
